Struct option_vec::OptionVec

pub struct OptionVec<T> { /* fields omitted */ }

An abstraction over Vec<Option<T>>

An element in an OptionVec<T> can be accessed by index and maintains its position when elements are removed from the container.

An element inserted into an OptionVec<T> will occupy the first available position in the container.

Methods

impl<T> OptionVec<T>

fn new() -> OptionVec<T>

Creates an empty OptionVec<T>.

fn with_capacity(n: usize) -> OptionVec<T>

Creates an empty OptionVec<T> with capacity for n elements.

fn inner(&self) -> &Vec<Option<T>>

Returns a borrowed reference to the internal Vec<Option<T>>.

fn inner_mut(&mut self) -> &mut Vec<Option<T>>

Returns a mutable reference to the internal Vec<Option<T>>.

Modification of this internal container is safe, but using methods such as Vec::insert or Vec::remove will invalidate existing indices.

fn capacity(&self) -> usize

Returns the allocated capacity for elements.

fn len(&self) -> usize

Returns the number of contained elements.

This operation is O(N), as all non-None elements must be individually counted.

fn is_empty(&self) -> bool

Returns whether the container is empty.

This operation is O(N) worst-case, as any elements must be searched for a non-None element.

fn insert(&mut self, t: T) -> usize

Inserts an element into the first available position, returning the destination position.

fn remove(&mut self, idx: usize) -> Option<T>

Removes an element from the given position, if one exists.

fn reserve(&mut self, n: usize)

Reserves capacity for at least n more elements.

fn reserve_exact(&mut self, n: usize)

Reserves capacity for exactly n more elements.

fn shrink_to_fit(&mut self)

Shrinks the allocation as much as possible.

Any trailing None elements will be truncated. None elements in internal positions are not removed, so as to maintain Some(_) element positions.

fn retain<F>(&mut self, f: F) where F: FnMut(&mut T) -> bool

Retains only elements specified by the predicate.

All elements e such that f(&mut e) returns false will be assigned to None.

fn pop_back(&mut self) -> Option<T>

Removes and returns the last occupied element.

fn pop_front(&mut self) -> Option<T>

Removes and returns the first occupied element.

fn clear(&mut self)

Removes all contained elements.

fn contains(&self, idx: usize) -> bool

Returns whether an element exists at the given index.

fn get(&self, idx: usize) -> Option<&T>

Returns an element at the given position.

fn get_mut(&mut self, idx: usize) -> Option<&mut T>

Returns a mutable reference to an element at the given position.

fn iter(&self) -> Iter<T>

Returns an iterator over contained elements.

fn iter_mut(&mut self) -> IterMut<T>

Returns an iterator over mutable references to contained elements.

fn into_enumerate(self) -> IntoEnumerate<T>

Returns an iterator over values with indices.

The elements yielded by this iterator will be (usize, T).

fn enumerate(&self) -> Enumerate<T>

Returns an iterator over references with indices.

The elements yielded by this iterator will be (usize, &T).

fn enumerate_mut(&mut self) -> EnumerateMut<T>

Returns an iterator over mutable references with indices.

The elements yielded by this iterator will be (usize, &mut T).